# On Redhat/Suse, use rsyslogd -v to determine the version