This method is much faster than the 'rpm -qa' dance that needs to run on every facter run. This fixes #204.

## Some benchmarks

Old version, `rpm -qa | grep ... `, 3 different runs:
```
3.142s
3.972s
2.740s
```

New version with `rsyslogd -v`, 3 different runs:
```
1.811s
1.778s
1.575s
```