@hahahn 2017-02-06
I believe the "rpm -qa" command which is still used as a fallback can be reduced to "rpm -q". The "a" is forcing a full rpm database scan which is resource-expensive call but without the "a", the query returns the same result very quickly.